# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NARA Academic Schedule & WhatsApp Reminder API",
        "description": "REST API for managing semesters, courses, lecturers, PICs, schedules, and automated WhatsApp reminders via Fonnte.",
        "contact": {
            "email": "admin@nara.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "NARA API Server"
        }
    ],
    "paths": {
...
        "/api/lecturers": {
            "get": {
                "tags": [
                    "Lecturers"
                ],
                "summary": "Get list of lecturers",
                "description": "Returns list of all lecturers",
                "operationId": "getLecturersList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Lecturers"
                ],
                "summary": "Create new lecturer",
                "description": "Creates a new lecturer record",
                "operationId": "storeLecturer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "gender"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ir. Rudy Ariyanto, S.T., M.Cs."
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "male",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "081234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lecturer created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/lecturers/{id}": {
            "get": {
                "tags": [
                    "Lecturers"
                ],
                "summary": "Get lecturer by ID",
                "description": "Returns single lecturer data",
                "operationId": "getLecturerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Lecturers"
                ],
                "summary": "Update existing lecturer",
                "description": "Updates a lecturer record",
                "operationId": "updateLecturer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Vivin Ayu Lestari, S.Pd., M.Kom."
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "female",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "081234567891"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lecturer updated successfully"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lecturers"
                ],
                "summary": "Delete lecturer",
                "description": "Deletes a lecturer record",
                "operationId": "deleteLecturer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lecturer deleted successfully"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
...
}